Move Views de wagtail_hooks.py para views.py
