+++ article/wagtail_hooks.py
@@ -1,6 +1,3 @@
-from django.contrib import messages
-from django.http import HttpResponseRedirect
-from django.shortcuts import redirect
 from django.urls import include, path
 from django.utils.translation import gettext as _
 from wagtail import hooks
@@ -9,106 +6,22 @@ from wagtail.contrib.modeladmin.options import (
     ModelAdminGroup,
     modeladmin_register,
 )
-from wagtail.contrib.modeladmin.views import CreateView, InspectView
 
+from article.views import (
+    ArticleAdminInspectView,
+    ArticleCreateView,
+    RelatedItemCreateView,
+    RequestArticleChangeCreateView,
+)
 from config.menu import get_menu_order
 
 from .button_helper import ArticleButtonHelper, RequestArticleChangeButtonHelper
-from .models import Article, RelatedItem, RequestArticleChange, choices
+from .models import Article, RelatedItem, RequestArticleChange
 from .permission_helper import ArticlePermissionHelper
 
 # from upload import exceptions as upload_exceptions
 # from upload.models import Package
 # from upload.tasks import get_or_create_package
-
-
-class ArticleCreateView(CreateView):
-    def form_valid(self, form):
-        self.object = form.save_all(self.request.user)
-        return HttpResponseRedirect(self.get_success_url())
-
-
-class RelatedItemCreateView(CreateView):
-    def form_valid(self, form):
-        self.object = form.save_all(self.request.user)
-        return HttpResponseRedirect(self.get_success_url())
-
-
-class RequestArticleChangeCreateView(CreateView):
-    def get_instance(self):
-        change_request_obj = super().get_instance()
-
-        article_id = self.request.GET.get("article_id")
-        if article_id:
-            article = Article.objects.get(pk=article_id)
-
-            if article:
-                change_request_obj.pid_v3 = article.pid_v3
-
-        return change_request_obj
-
-    # # FIXME
-    # def form_valid(self, form):
-    #     pid_v3 = self.request.POST['pid_v3']
-
-    #     try:
-    #         package_id = get_or_create_package(
-    #             pid_v3=pid_v3,
-    #             user_id=self.request.user.id
-    #         )
-    #     except upload_exceptions.XMLUriIsUnavailableError as e:
-    #         messages.error(
-    #             self.request,
-    #             _('It was not possible to submit the request. XML Uri is unavailable: %s.') % e.uri,
-    #         )
-    #         return redirect(self.request.META.get('HTTP_REFERER'))
-    #     except upload_exceptions.PIDv3DoesNotExistInSiteDatabase:
-    #         messages.error(
-    #             self.request,
-    #             _('It was not possible to submit the request. PIDv3 does not exist in the site database.'),
-    #         )
-    #         return redirect(self.request.META.get('HTTP_REFERER'))
-    #     except upload_exceptions.SiteDatabaseIsUnavailableError:
-    #         messages.error(
-    #             self.request,
-    #             _('It was not possible to submit the request. Site database is unavailable.'),
-    #         )
-    #         return redirect(self.request.META.get('HTTP_REFERER'))
-
-    #     article = Package.objects.get(pk=package_id).article
-
-    #     change_request_obj = form.save_all(self.request.user, article)
-
-    #     if change_request_obj.change_type == choices.RCT_ERRATUM:
-    #         article.status =  choices.AS_REQUIRE_ERRATUM
-    #     elif change_request_obj.change_type ==  choices.RCT_UPDATE:
-    #         article.status = choices.AS_REQUIRE_UPDATE
-
-    #     article.save()
-
-    #     messages.success(
-    #         self.request,
-    #         _('Change request submitted with success.')
-    #     )
-    #     return HttpResponseRedirect(self.get_success_url())
-
-
-class ArticleAdminInspectView(InspectView):
-    def get_context_data(self):
-        data = {
-            "status": self.instance.status,
-            "packages": self.instance.package_set.all(),
-        }
-
-        if self.instance.status in (
-            choices.AS_REQUIRE_UPDATE,
-            choices.AS_REQUIRE_ERRATUM,
-        ):
-            data["requested_changes"] = []
-            for rac in self.instance.requestarticlechange_set.all():
-                data["requested_changes"].append(rac)
-
-        return super().get_context_data(**data)
 
 
 class ArticleModelAdmin(ModelAdmin):
@@ -123,25 +36,29 @@ class ArticleModelAdmin(ModelAdmin):
     menu_order = get_menu_order("article")
     add_to_settings_menu = False
     exclude_from_explorer = False
+    list_per_page = 20
 
     list_display = (
         "__str__",
-        "pid_v3",
-        # "pid_v2",
-        # "doi_list",
-        # "aop_pid",
-        # "article_type",
         "status",
+        "display_sections",
+        "order",
+        "fpage",
+        "position",
+        "first_publication_date",
         "created",
         "updated",
         # "updated_by",
     )
     list_filter = ("status",)
     search_fields = (
+        "sps_pkg__sps_pkg_name",
         "pid_v3",
         "issue__publication_year",
-        "journal__title",
-        "sps_pkg__sps_pkg_name",
+        "journal__official_journal__title",
+        "journal__official_journal__issn_print",
+        "journal__official_journal__issn_electronic",
+        "title_with_lang__text",
     )
     inspect_view_fields = (
         "created",
@@ -154,6 +71,7 @@ class ArticleModelAdmin(ModelAdmin):
         "doi_with_lang",
         "article_type",
         "status",
+        "issue",
         # "author",
         # "title_with_lang",
         "elocation_id",
@@ -235,16 +153,15 @@ class RequestArticleChangeModelAdmin(ModelAdmin):
 class ArticleModelAdminGroup(ModelAdminGroup):
     menu_label = _("Articles")
     menu_icon = "folder-open-inverse"
-    # menu_order = get_menu_order("article")
-    menu_order = 400
+    menu_order = get_menu_order("article")
     items = (
         ArticleModelAdmin,
         # RelatedItemModelAdmin,
         # RequestArticleChangeModelAdmin,
+        # ApprovedArticleModelAdmin,
     )
 
 
-# modeladmin_register(ArticleModelAdminGroup)
 modeladmin_register(ArticleModelAdmin)
